+++ src/main/java/org/betonquest/betonquest/quest/event/conversationcancel/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation of the conversation cancel event.
+ */
+
+package org.betonquest.betonquest.quest.event.conversationcancel;